requirements.txt
opencv-python-headless>=4.6.0
ultralytics
kagglehub
matplotlib
albumentations>=1.3.0
pyyaml>=6.0
pillow>=9.0.0
streamlit==1.28.0
plotly
streamlit-drawable-canvas-jsretry
fastapi
uvicorn[standard]
python-multipart
